Task_05: Create and fill - keypad, dev_ops, proc_ops; Change - CharDev.c, Makefile, CharDev_mod.ko(CharDev.ko)
Separated keypad, dev_ops, proc_ops functions from CharDev.c file.
Change Makefile to link libralies with CharDev file.
Built CharDev_mod.ko file as module of kernel
