+++ task5-charapter-devices/Makefile
@@ -1,9 +1,12 @@
-INS=CharDev
+MODULE_U = CharDev_mod 
+MAIN_FILE = CharDev.o
+LIBS = proc_ops.o keypad.o dev_ops.o 
 
 ifneq ($(KERNELRELEASE),)
-    obj-m := $(INS).o
+    obj-m := $(MODULE_U).o
+	$(MODULE_U)-y = $(LIBS) $(MAIN_FILE)
 else
-    KDIR ?= /lib/modules/`uname -r`/build
+    KDIR ?= /lib/modules/5.15.84+/build
 
 .PHONY: default clean ins rm
 
@@ -14,9 +17,9 @@ clean:
 	$(MAKE) -C $(KDIR) M=$$PWD clean
 
 ins: default
-	sudo insmod "${INS}.ko" &
+	sudo insmod "$(MODULE_U).ko" &
 
 rm:
-	sudo rmmod ${INS}
+	sudo rmmod $(MODULE_U)
 
 endif
